+++ src/components/list/list-header.ts
@@ -2,6 +2,7 @@ import { LitElement, html } from 'lit';
 
 import { styles } from './themes/header.base.css.js';
 import { all } from './themes/header.js';
+import { styles as shared } from './themes/shared/header/list-header.common.css.js';
 import { themes } from '../../theming/theming-decorator.js';
 import { registerComponent } from '../common/definitions/register.js';
 
@@ -15,7 +16,7 @@ import { registerComponent } from '../common/definitions/register.js';
 @themes(all)
 export default class IgcListHeaderComponent extends LitElement {
   public static readonly tagName = 'igc-list-header';
-  public static override styles = styles;
+  public static override styles = [styles, shared];
 
   public static register() {
     registerComponent(this);